work around priority queue overflow
The priority queue may become fragmented after a significant number of
insertions and deletes so that all slots in all groups are being used,
but the last external group is nearly empty (counting elements).
In that case we merge the whole last external group into a temporary
ext_merger and swap this overflow merger and the (now empty) last
external merger afterwards. Increasing the number of groups beyond 4 is
not possible without significant changes elsewhere.


git-svn-id: svn://svn.code.sf.net/p/stxxl/code/trunk@3189 4e380d45-d1fd-0310-85a7-d18ec86df0ad